Add functionality for star button clicks.
On click star icon classes (bi-star and bit-star-fill) are both toggled.
Event click handler checks logic if story is favorited or not and adds or removes appropriately.
On page load, each story is check if it is favorited and stars begin on the correct status of filled or not filled.

TODO: make a separate section that can be navigated to that only contains favorited stories.